backend/requirements.txt
asgiref==3.7.2
django-cors-headers==4.3.1
daphne==4.0.0
Django==4.2.7
django-filter==23.4
django-gisserver==1.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
uwsgi==2.0.23
tzdata==2023.3
